WIP - Refactoring graphics pipelines
